Reformatted Build.scala
