+++ kifi-backend/project/plugins.sbt
@@ -10,3 +10,5 @@ addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.2.2")
 addSbtPlugin("com.typesafe.sbt" % "sbt-scalariform" % "1.3.0")
 
 scalacOptions ++= Seq("-unchecked", "-deprecation", "-feature")
+
+addSbtPlugin("net.virtual-void" % "sbt-dependency-graph" % "0.7.4")
